made some changes and fixed assignment son
