+++ mykingapp/ScheduleViewController.swift
@@ -74,7 +74,7 @@ class ScheduleViewController: UITableViewController {
         
         super.viewDidLoad()
         
-        scheduleData = ScheduleData.getScheduleData(fName: "Ryan", lName: "Heaton", grade: "21")
+        scheduleData = ScheduleData.getScheduleData(fName: "Ryan", lName: "Heaton", grade: 21)
         print("Data that I have now: \(scheduleData!.name)")
 
         // Uncomment the following line to preserve selection between presentations
